+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { GifCard } from "./gif-card/gif-card";
+export { GifGrid } from "./gif-grid/gif-grid";
+export { SearchBar } from "./search-bar/search-bar";
+export { Tofy } from "./tofy/tofy";
